requirements.txt
Flask==3.0.2
gunicorn==21.2.0
kafka-python==2.1.0
msgspec==0.18.6
redis==5.0.3
requests==2.31.0
asyncio==3.4.3
pytest==8.3.5
pytest-asyncio==0.25.3